requirements.txt
annotated-types==0.7.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pdfminer.six==20240706
preshed==3.0.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-docx==1.1.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
setuptools==71.0.4
shellingham==1.5.4
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
